Move CreateResponseAsync to extension methods; avoid duplicate reflection
