+++ FastPass/FastPass.API/Functions/FhirServiceProxyFunction.cs
@@ -34,65 +34,71 @@ namespace FastPass.API.Functions
             try
             {
                 var idFromUri = req.Url.LocalPath.Split('/').Last().Replace("Patient", "");
+
                 IList<Patient> returnedPatients = null;
+
                 if (!string.IsNullOrWhiteSpace(idFromUri))
                 {
                     var filter = new PatientFilters
                     {
                         FhirId = idFromUri
                     };
-                    _logger.LogInformation("FhirService::GetPatient called for {patientId}.", idFromUri);
+
+                    _logger.LogInformation($"FhirService::GetPatient called for {idFromUri}.");
 
                     returnedPatients = await _fhirService.GetPatientsAsync(filter);
                 }
                 else
                 {
                     _logger.LogInformation("FhirService::GetPatient called for all patients.");
+
                     returnedPatients = await _fhirService.GetPatientsAsync();
                 }
 
-                var resp = req.CreateResponse(HttpStatusCode.OK);
-                await resp.WriteAsJsonAsync(returnedPatients);
-                return resp;
+                var json = JsonConvert.SerializeObject(returnedPatients);
+
+                return await req.CreateResponseAsync(HttpStatusCode.OK, json);
             }
             catch (Exception ex)
             {
                 var msg = $"FhirService::GetPatient failed. Detail: {ex}";
+
                 _logger.LogError("FhirService::GetPatient failed. Detail: {MessageException}", msg);
 
-                var err = req.CreateResponse(HttpStatusCode.BadRequest);
-                await err.WriteStringAsync(msg);
-                return err;
+                return await req.CreateResponseAsync(HttpStatusCode.BadRequest, msg);
             }
-
-
         }
 
         [Function("AddPatient")]
         public async Task<HttpResponseData> AddPatient([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = null)] HttpRequestData req)
         {
+            var fhirFx = $"{nameof(_fhirService)}::{nameof(_fhirService.CreatePatientAsync)}";
+
             try
             {
                 var bodyString = await req.GetBodyString();
+
                 var parser = new FhirJsonParser();
+
                 var patient = parser.Parse<Patient>(bodyString);
 
-                _logger.LogInformation("Calling FhirService::AddPatient");
-                var returnedPatient = await _fhirService.CreatePatientAsync(patient);
-                _logger.LogInformation("FhirService::AddPatient succeeded. {patientId} created.", returnedPatient.Id);
+                _logger.LogInformation($"Calling {fhirFx}");
 
-                var resp = req.CreateResponse(HttpStatusCode.OK);
-                await resp.WriteAsJsonAsync(returnedPatient);
-                return resp;
+                var returnedPatient = await _fhirService.CreatePatientAsync(patient);
+
+                _logger.LogInformation($"{fhirFx} succeeded. {returnedPatient.Id} created.");
+
+                var json = JsonConvert.SerializeObject(returnedPatient);
+
+                return await req.CreateResponseAsync(HttpStatusCode.OK, json);
             }
             catch (Exception ex)
             {
-                var msg = $"FhirService::AddPatient failed. Detail: {ex}";
+                var msg = $"{fhirFx} failed. Detail: {ex}";
+
                 _logger.LogError(msg);
 
-                var err = req.CreateResponse(HttpStatusCode.BadRequest);
-                await err.WriteStringAsync(msg);
-                return err;
+                return await req.CreateResponseAsync(HttpStatusCode.BadRequest, msg);
             }
         }
 
@@ -128,28 +134,33 @@ namespace FastPass.API.Functions
         [Function("AddObservation")]
         public async Task<HttpResponseData> AddObservation([HttpTrigger(AuthorizationLevel.Anonymous, "post", Route = null)] HttpRequestData req)
         {
+            var fhirFx = $"{nameof(_fhirService)}::{nameof(_fhirService.CreateObservationAsync)}";
+
             try
             {
                 var bodyString = await req.GetBodyString();
+
                 var parser = new FhirJsonParser();
+
                 var observation = parser.Parse<Observation>(bodyString);
 
-                _logger.LogInformation($"Calling {nameof(_fhirService)}::{nameof(_fhirService.CreateObservationAsync)}");
-                var returnedObservation = await _fhirService.CreateObservationAsync(observation);
-                _logger.LogInformation($"{nameof(_fhirService)}::{nameof(_fhirService.CreateObservationAsync)} succeeded. {returnedObservation.Id} created.");
+                _logger.LogInformation($"Calling {fhirFx}");
 
-                var resp = req.CreateResponse(HttpStatusCode.OK);
-                await resp.WriteAsJsonAsync(returnedObservation);
-                return resp;
+                var returnedObservation = await _fhirService.CreateObservationAsync(observation);
+
+                _logger.LogInformation($"{fhirFx} succeeded. {returnedObservation.Id} created.");
+
+                var json = JsonConvert.SerializeObject(returnedObservation);
+                
+                return await req.CreateResponseAsync(HttpStatusCode.OK, json);
             }
             catch (Exception ex)
             {
-                var msg = $"{nameof(_fhirService)}::{nameof(_fhirService.CreateObservationAsync)} failed. Detail: {ex}";
+                var msg = $"{fhirFx} failed. Detail: {ex}";
+
                 _logger.LogError(msg);
 
-                var err = req.CreateResponse(HttpStatusCode.BadRequest);
-                await err.WriteStringAsync(msg);
-                return err;
+                return await req.CreateResponseAsync(HttpStatusCode.BadRequest, msg);
             }
         }
     }
