+++ FastPass/FastPass.API/Functions/TextAnalyticsServiceProxyFunction.cs
@@ -43,12 +43,12 @@ public class TextAnalyticsServiceProxyFunction
             {
                 var msg = $"No values returned from Text Analytics.";
                 _logger.LogWarning(msg);
-                return await CreateResponseAsync(req, HttpStatusCode.BadRequest, msg);
+                return await req.CreateResponseAsync(HttpStatusCode.BadRequest, msg);
             }
 
             var results = healthOps.GetValues();
             var fhirResults = results.SelectMany(p => p.Select(d => d.FhirBundle)).FirstOrDefault();
-            return await CreateResponseAsync(req, HttpStatusCode.OK, JsonConvert.SerializeObject(fhirResults));
+            return await req.CreateResponseAsync(HttpStatusCode.OK, JsonConvert.SerializeObject(fhirResults));
 
         }
         catch (Exception ex)
@@ -56,20 +56,7 @@ public class TextAnalyticsServiceProxyFunction
             var msg = $"TextAnalytics exception caught. Detail: {ex}";
             _logger.LogError(msg);
 
-            return await CreateResponseAsync(req, HttpStatusCode.BadRequest, msg);
+            return await req.CreateResponseAsync(HttpStatusCode.BadRequest, msg);
         }
-    }
-
-
-    private async Task<HttpResponseData> CreateResponseAsync(HttpRequestData req, HttpStatusCode responseCode, string body)
-    {
-        var resp = req.CreateResponse(responseCode);
-
-        await resp.WriteStringAsync(body);
-
-        if (responseCode.IsSuccessful())
-            resp.Headers.Add("Content-Type", "application/json");
-
-        return resp;
     }
 }
